requirements.txt
requests
hypy_utils>=1.0.15
py-cpuinfo
tqdm
